requirements.txt
Django==5.0.1
python-dotenv==1.0.0
openai==1.6.1
python-decouple==3.8
whitenoise==6.5.0
gunicorn==21.2.0